Move magic_loadpy from GUI-specific magics to core magics.
